[Feature] add redux and restructure project (#2)
