+++ package.json
@@ -10,7 +10,11 @@
     "react": "16.5.0",
     "react-native": "0.57.1",
     "react-native-linear-gradient": "^2.4.2",
-    "react-navigation": "^2.17.0"
+    "react-navigation": "^2.17.0",
+    "react-navigation-redux-helpers": "^2.0.6",
+    "react-redux": "^5.0.7",
+    "redux": "^4.0.0",
+    "redux-logger": "^3.0.6"
   },
   "devDependencies": {
     "@babel/runtime": "^7.1.2",
