+++ src/navigation/reducers/index.js
@@ -1,0 +1,4 @@
+import {createNavigationReducer} from 'react-navigation-redux-helpers';
+import AppNavigator from '../navigators'
+
+export default createNavigationReducer(AppNavigator);
