+++ src/reducers/index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from "redux";
+import navigationData from "../navigation/reducers";
+
+export default combineReducers({
+    navigationData
+});
